requirements.txt
frappe
websocket_client
